refactor(plugin-js-packages): organise package managers in separate folders
