+++ packages/plugin-js-packages/src/lib/package-managers/constants.ts
@@ -1,0 +1,2 @@
+export const COMMON_AUDIT_ARGS = ['audit', '--json'];
+export const COMMON_OUTDATED_ARGS = ['outdated', '--json'];
