+++ packages/plugin-js-packages/src/lib/package-managers/index.ts
@@ -1,0 +1,2 @@
+export { packageManagers } from './package-managers';
+export { PackageManager } from './types';
