+++ packages/plugin-js-packages/src/lib/package-managers/package-managers.ts
@@ -1,0 +1,13 @@
+import { PackageManagerId } from '../config';
+import { npmPackageManager } from './npm/npm';
+import { pnpmPackageManager } from './pnpm/pnpm';
+import { PackageManager } from './types';
+import { yarnv1PackageManager } from './yarn-classic/yarn-classic';
+import { yarnv2PackageManager } from './yarn-modern/yarn-modern';
+
+export const packageManagers: Record<PackageManagerId, PackageManager> = {
+  npm: npmPackageManager,
+  'yarn-classic': yarnv1PackageManager,
+  'yarn-modern': yarnv2PackageManager,
+  pnpm: pnpmPackageManager,
+};
