+++ packages/plugin-js-packages/src/lib/package-managers/pnpm/audit-result.ts
@@ -1,0 +1,52 @@
+import { AuditResult, Vulnerability } from '../../runner/audit/types';
+import { getVulnerabilitiesTotal } from '../../runner/audit/utils';
+import { PnpmAuditResultJson } from './types';
+
+export function pnpmToAuditResult(output: string): AuditResult {
+  const pnpmResult = JSON.parse(output) as PnpmAuditResultJson;
+
+  const vulnerabilities = Object.values(pnpmResult.advisories).map(
+    ({
+      module_name: name,
+      id,
+      title,
+      url,
+      severity,
+      vulnerable_versions: versionRange,
+      recommendation: fixInformation,
+      findings,
+    }): Vulnerability => {
+      const path = findings[0]?.paths[0];
+
+      return {
+        name,
+        id,
+        title,
+        url,
+        severity,
+        versionRange,
+        directDependency: path == null ? true : pnpmToDirectDependency(path),
+        fixInformation,
+      };
+    },
+  );
+
+  return {
+    vulnerabilities,
+    summary: {
+      ...pnpmResult.metadata.vulnerabilities,
+      total: getVulnerabilitiesTotal(pnpmResult.metadata.vulnerabilities),
+    },
+  };
+}
+
+export function pnpmToDirectDependency(path: string): string | true {
+  // the format is ". > <direct dependency>@<version> > ... > <current dependency>@<version>"
+  const deps = path.split(' > ').slice(1);
+
+  if (deps.length <= 1) {
+    return true;
+  }
+
+  return deps[0]?.split('@')[0] ?? true;
+}
