+++ packages/plugin-js-packages/src/lib/package-managers/yarn-modern/outdated-result.ts
@@ -1,0 +1,13 @@
+import { OutdatedResult } from '../../runner/outdated/types';
+import { Yarnv2OutdatedResultJson } from './types';
+
+export function yarnv2ToOutdatedResult(output: string): OutdatedResult {
+  const npmOutdated = JSON.parse(output) as Yarnv2OutdatedResultJson;
+
+  return npmOutdated.map(({ name, current, latest, type }) => ({
+    name,
+    current,
+    latest,
+    type,
+  }));
+}
