+++ packages/plugin-js-packages/src/lib/runner/audit/utils.ts
@@ -1,0 +1,7 @@
+import { PackageAuditLevel } from '../../config';
+
+export function getVulnerabilitiesTotal(
+  summary: Record<PackageAuditLevel, number>,
+): number {
+  return Object.values(summary).reduce((acc, value) => acc + value, 0);
+}
